+++ typescript/packages/core/src/types/Chain.ts
@@ -2,7 +2,7 @@
  * @param type - "evm" or "solana", extend this union as needed (e.g., "sui")
  * @param id - Chain ID, optional for EVM
  */
-export type Chain = EvmChain | SolanaChain | AptosChain | ChromiaChain;
+export type Chain = EvmChain | SolanaChain | AptosChain | ChromiaChain | FuelChain;
 
 export type EvmChain = {
     type: "evm";
@@ -19,4 +19,8 @@ export type AptosChain = {
 
 export type ChromiaChain = {
     type: "chromia";
+};
+
+export type FuelChain = {
+    type: "fuel";
 };