+++ typescript/packages/plugins/mode-governance/src/index.ts
@@ -1,0 +1,19 @@
+import { Chain, PluginBase } from "@goat-sdk/core";
+import { EVMWalletClient } from "@goat-sdk/wallet-evm";
+import { mode } from "viem/chains";
+import { ModeGovernanceService } from "./mode-governance.service";
+import { GetBalanceParameters, GetStakeInfoParameters, StakeParameters } from "./parameters";
+
+const SUPPORTED_CHAINS = [mode];
+
+export class ModeGovernancePlugin extends PluginBase<EVMWalletClient> {
+    constructor() {
+        super("mode-governance", [new ModeGovernanceService()]);
+    }
+
+    supportsChain = (chain: Chain) => chain.type === "evm" && SUPPORTED_CHAINS.some((c) => c.id === chain.id);
+}
+
+export const modeGovernance = () => new ModeGovernancePlugin();
+
+export * from "./constants";