+++ typescript/packages/plugins/meteora/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @goat-sdk/plugin-meteora
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [279dfcd]
+  - @goat-sdk/core@0.4.2
+  - @goat-sdk/wallet-solana@0.2.6
+
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [6c03f61]
+  - @goat-sdk/core@0.4.1
+  - @goat-sdk/wallet-solana@0.2.5
+
 ## 0.2.4
 
 ### Patch Changes