+++ typescript/packages/plugins/polymarket/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @goat-sdk/plugin-polymarket
 
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [279dfcd]
+  - @goat-sdk/core@0.4.2
+  - @goat-sdk/wallet-evm@0.2.2
+
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies [6c03f61]
+  - @goat-sdk/core@0.4.1
+  - @goat-sdk/wallet-evm@0.2.1
+
+## 0.3.2
+
+### Patch Changes
+
+- 1bc1d42: Minor changes
+
 ## 0.3.1
 
 ### Patch Changes