+++ typescript/packages/wallets/crossmint/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @goat-sdk/crossmint
 
+## 0.3.8
+
+### Patch Changes
+
+- Updated dependencies [279dfcd]
+  - @goat-sdk/core@0.4.2
+  - @goat-sdk/wallet-evm@0.2.2
+  - @goat-sdk/wallet-solana@0.2.6
+
+## 0.3.7
+
+### Patch Changes
+
+- Updated dependencies [6c03f61]
+  - @goat-sdk/core@0.4.1
+  - @goat-sdk/wallet-evm@0.2.1
+  - @goat-sdk/wallet-solana@0.2.5
+
+## 0.3.6
+
+### Patch Changes
+
+- 1bc1d42: Add MODE network
+
 ## 0.3.5
 
 ### Patch Changes